requirements.txt
python-decouple
telethon
pytgcalls==3.0.0.dev19
youtube-dl
pafy
